[#78] domain interface 분리
