+++ src/api.rs
@@ -3,6 +3,7 @@ pub(crate) mod background;
 pub(crate) mod context;
 pub(crate) mod db;
 pub(crate) mod docker;
+pub(crate) mod domain;
 pub(crate) mod repositories;
 pub(crate) mod routes;
 pub(crate) mod types;
