+++ src/domain/job.rs
@@ -1,0 +1,11 @@
+use crate::{
+    db::entities,
+    repositories::{CreateJobParams, ListJobsParams, PatchJobParams},
+};
+
+#[async_trait::async_trait]
+pub trait JobRepository {
+    async fn list_jobs(&self, params: ListJobsParams) -> anyhow::Result<Vec<entities::job::Model>>;
+    async fn create_job(&self, params: CreateJobParams) -> anyhow::Result<i64>;
+    async fn patch_job(&self, params: PatchJobParams) -> anyhow::Result<()>;
+}
